Refactor backend

+++ backend/src/main/java/com/bounswe2024group10/Tradeverse/repository/DislikeRepository.java
@@ -14,5 +14,5 @@ public interface DislikeRepository extends JpaRepository<Dislike, Long> {
     Boolean existsByUsernameAndPostID(String username, Long postID);
     List<Dislike> findByUsername(String username);
     List<Dislike> findByPostID(Long postID);
-    Long countByPostID(Long postID);
+    int countByPostID(Long postID);
 }
